pyproject.toml
[tool.poetry]
name = "stepn"
version = "0.1.0"
description = ""
authors = ["EinNarr"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
python-dotenv = "^0.20.0"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
